requirements.txt
# Flask
Flask==2.3.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7

# Astronomy
swisseph==2.10.3
ephem==4.1.5
zhdate==0.2
skyfield==1.45
jplephem==2.18
numpy==1.24.3
pandas==2.0.3
pytz==2023.3

# RAG Dependencies
jieba==0.42.1
requests==2.31.0
